Detect careless responses

A function to detect careless survey responses using the careless package. The function has the improvement to first remove speeders, because they negatively affect the metrics used in the careless package.

careless_indices(
  dat,
  speeder_analysis = "median/2",
  likert_vector = NULL,
  duration_column = "Duration (in seconds)",
  id_column = "ResponseId",
  antonyms = TRUE,
  synonyms = TRUE
)

Arguments

dat
A data.frame that contains the survey data.

speeder_analysis
A criterion to test against speeders. The speeder_analysis parameter can be one of the following: median, median/2 (default), median/3, or a numeric value that is directly used as a threshold. Using FALSE will ignore speeder analysis.

likert_vector
A vector of indices that should be analyzed.

duration_column
A string that contains the name of the column that has the duration of survey responses. As a default we use Duration (in seconds), which is used by Qualtrix.

id_column
A string that contains the name of a id column. As a default we use "ResponseId", which is used by Qualtrix.

antonyms
Whether or not to test for psychometric antonyms (default: T)

synonyms
Whether or not to test for psychometric synonyms (default: T)

Value

Returns an updated dataframe containing indices